Tested that driver works by mmap-ing to BAR0
Done with python3. Open /dev/mem as a file, then mmap that file at the
offset given by the output of the `lspci -vv` command for the FPGA. If
everything worked correctly, the 32 bits of 0xDEADBEEF should be
stored in little-endian order (0xef, 0xbe, 0xad, 0xde), and printed
out as such.