Change makefile so it is usable with buildroot
Buildroot does not know what to do if there is a kernel module that is
built with a main file that is named differently than the kernel
module. So, fpga_char_main -> fpga_char, and then when building on the
host side (meaning building on a computer for that computer) a
circular dependency is introduced that is automatically dropped. This
is fine, because this is a simple project, file-wise.