Add co-processing thread that cleans up the provided virtine addrs
This is HEAVILY inspired by and somewhat stolen from QEMU's EDU PCI
device.

How this works:
processing_thread is the thread of execution that actually cleans up
the virtines. The lock is used to ensure that the various accesses to
locations are well-ordered. The condition is used to ensure that the
co-processor thread waits (scheduler wait) until the doorbell is rung.
When the doorbell is "rung", the condition is signalled, which allows
the co-processor thread to exit the inner while-loop and move on with
execution. The first thing to do is to set the various values of the
FPGA. Once the doorbell and is_card_processing variables are set, any
kind of operation can happen. Lastly, the PCI device raises an
interrupt that the CPU/kernel module will have to handle.