Remove unneeded release_device function
The purpose of the function was to give a nice way to clean up the PCI
device initialization. But, with the fpga_probe function using a goto
chain for clean-up in the case of failure, this was rendered useless.
So, all of the clean-up steps in release_device were simply moved into
fpga_remove, which is called when the PCI device is removed from the system.