Ensure that if pop_head returns an invalid hwaddr, do NOT clean it
pop_head returns a hwaddr (an element from the queue), and if it an
element that has been popped already (meaning the queue is now empty),
then an invalid hwaddr is returned (0). When this is typecast to a
hwaddr*, it can lead to a null dereference in the memcpy. To prevent
this, we need to skip the part of the loop where the memcpy happens,
but we still NEED to unlock the mutex. Hence the need for a goto to
the end of the loop.