Reading addresses from queues uses static var to split return
This is necessary to do because I do not have support for
reading/writing 64-bit values from/to the FPGA device. This appears to
be a kernel limitation, as those functions are simply not defined.
To get around this, I must perform 2 32-bit reads/writes. When reading
the hardware addresses of the cleaned virtines from the FPGA, I must
split the address in half essentially. By using a static variable, I
can assign to it when I pop the address and then re-use it (shifting
bits down by 32) on the next call to get the next 32 bits.