Write test program to insert a "virtine" using pwrite command
TODO: Switch to using _JUST_ the write syscall instead of pwrite. You
should be able to write addresses and the system manages to figure it
out for you.

This program allocates an integer on the heap and passes the pointer
to the heap memory to the virtine function to write _as if_ it were a
virtine.