Use static split_write var to store and write 2 32-bit nums as 64-bits
Similar to the reason the static_read variable is used in the
virtine_fpga_mmio_read function to read a 64-bit value using two
32-bit ioread32 calls, we need a static variable to store the gotten
intermediate value from the kernel. When we receive all 8 bytes, we
use bitwise operations to reconstruct the 8 byte integer from the two
4 byte integers.

Then, we lock the queue and insert the integer to the tail, before unlocking.