Clean up virtine by using DMA write
QEMU has to run through several steps to be able to write to a linux
user-space memory location running in the guest, and you wanted to
de-reference some pointer in that program you would need to:
    1.) Convert the guest virtual address into a guest physical
    address (usually done in the guest)
    2.) Pass that address to the PCI device (e.g. by writing to some
    register)
    3.) Convert the guest physical address to a host virtual
    address (this is handled by QEMU's dma apis).
This occurs because a userspace malloc could be scattered all over
guest physical memory or e.g. swapped out by the guest kernel.
Credit to: a1xndr from #qemu on irc.oftc.net for helping me figure
this out.

To simplify things, the pci_dma_write command is used instead.